move components to the components folder
