+++ .gitignore
@@ -44,7 +44,7 @@ node_modules
 
 # next.js
 .next/
-/out/
+out/
 
 # production
 /build
