+++ frontend/components/Customer/ProgramIsActive/index.js
@@ -13,11 +13,10 @@ const ProgramIsActive = ({
   purchaseWithTokens,
   buyWithCCLoader,
   buyWithTokensLoader,
-  managerContract,
 }) => (
   <>
     <div className="grid grid-cols-3 gap-10 mt-10">
-      <AccountDetails customerUuid={customerUuid} ftMetadata={ftMetadata} customerBalance={customerBalance} managerContract={managerContract}/>
+      <AccountDetails customerUuid={customerUuid} ftMetadata={ftMetadata} customerBalance={customerBalance} />
       <Rules ftMetadata={ftMetadata} product={product} />
     </div>
