+++ frontend/components/Customer/Rules/index.js
