--- frontend/pages/customer/Customer/Header/index.js
@@ -1,5 +1,0 @@
-import HeaderSection from '../../../../components/HeaderSection';
-
-const Header = () => <HeaderSection />;
-
-export default Header;
